Change the data type for the number of screen divisions in the Rohde&Schwarz/Hameg oscilloscope driver.
The oscilloscope configuration data structure is now accessed
both in read and write mode and some member variables are
written at runtime. The member variables' data types should
therefore match those used when writing them, otherwise in the
case, for example, of fixed-width integer types, those variables
might be written past their boundaries, leading to data
corruption.

In particular this commit fixes a recently introduced bug
which causes the number of vertical (y) divisions to be
always set to zero (previous member variable written past its
boundaries, i.e. uint8_t as int).

Such minor bug has been introduced with the following commit:

commit e56bb53105b7f92387713aad4946a34efa1957a3
Date:   Tue Nov 20 23:24:07 2018 +0100

---
 src/hardware/rohde-schwarz-hameg/protocol.h |    4 ++--
 1 file changed, 2 insertions(+), 2 deletions(-)

+++ src/hardware/rohde-schwarz-hameg/protocol.h
@@ -159,8 +159,8 @@ struct scope_config {
 	const uint64_t (*vscale)[][2];
 	const uint8_t num_vscale;
 
-	uint8_t num_xdivs;
-	const uint8_t num_ydivs;
+	unsigned int num_xdivs;
+	const unsigned int num_ydivs;
 
 	const char *(*scpi_dialect)[];
 };
